backend/pyproject.toml
[project]
name = "terview-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pydantic[email]>=2.11.9",
    "requests>=2.32.3",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "pyyaml>=6.0.2",
]
